+++ components/app/style/index.ts
@@ -1,5 +1,5 @@
 import type { FullToken, GenerateStyle, GetDefaultToken } from '../../theme/internal';
-import { genComponentStyleHook } from '../../theme/internal';
+import { genStyleHooks } from '../../theme/internal';
 
 export type ComponentToken = {};
 
@@ -21,4 +21,8 @@ const genBaseStyle: GenerateStyle<AppToken> = (token) => {
 export const prepareComponentToken: GetDefaultToken<'App'> = () => ({});
 
 // ============================== Export ==============================
-export default genComponentStyleHook<'App'>('App', genBaseStyle, prepareComponentToken);
+export const [useStyle, useCSSVar] = genStyleHooks<'App'>(
+  'App',
+  genBaseStyle,
+  prepareComponentToken,
+);
