--- components/back-top/style/cssVar.ts
@@ -1,4 +1,0 @@
-import { prepareComponentToken } from '.';
-import { genCSSVarRegister } from '../../theme/internal';
-
-export default genCSSVarRegister<'BackTop'>('BackTop', prepareComponentToken);
